+++ Analytics/Providers/Amplitude/AmplitudeProvider.m
@@ -43,13 +43,13 @@
 #pragma mark - Analytics API
 
 
-- (void)identify:(NSString *)userId traits:(NSDictionary *)traits context:(NSDictionary *)context
+- (void)identify:(NSString *)userId traits:(NSDictionary *)traits options:(NSDictionary *)options
 {
     [Amplitude setUserId:userId];
     [Amplitude setGlobalUserProperties:traits];
 }
 
-- (void)track:(NSString *)event properties:(NSDictionary *)properties context:(NSDictionary *)context
+- (void)track:(NSString *)event properties:(NSDictionary *)properties options:(NSDictionary *)options
 {
     [Amplitude logEvent:event withCustomProperties:properties];
 
@@ -60,10 +60,10 @@
     }
 }
 
-- (void)screen:(NSString *)screenTitle properties:(NSDictionary *)properties context:(NSDictionary *)context
+- (void)screen:(NSString *)screenTitle properties:(NSDictionary *)properties options:(NSDictionary *)options
 {
     // No explicit support for screens, so we'll track an event instead.
-    [self track:screenTitle properties:properties context:context];
+    [self track:screenTitle properties:properties options:options];
 }
 
 @end